Rename package to avoid conflicts with Strange (part 2)
