+++ src/main/java/org/redfx/strangefx/ui/Main.java
@@ -32,7 +32,7 @@
  */
 package org.redfx.strangefx.ui;
 
-import org.redfx.strange.simulator.Model;
+import org.redfx.strangefx.simulator.Model;
 import org.redfx.strange.gate.*;
 import org.redfx.strange.simulator.local.LocalSimulator;
 import javafx.application.Application;
