+++ src/main/java/org/redfx/strangefx/ui/QubitBoard.java
@@ -32,7 +32,7 @@
  */
 package org.redfx.strangefx.ui;
 
-import org.redfx.strange.simulator.Model;
+import org.redfx.strangefx.simulator.Model;
 import org.redfx.strange.ui.render.*;
 import javafx.beans.Observable;
 import javafx.collections.FXCollections;
